+++ lib/lightning/jobs.ex
@@ -194,7 +194,12 @@ defmodule Lightning.Jobs do
 
   """
   def delete_job(%Job{} = job) do
-    Repo.delete(job)
+    job
+    |> Ecto.Changeset.change()
+    |> Ecto.Changeset.foreign_key_constraint(:trigger_id,
+      message: "This job is associated with downstream jobs"
+    )
+    |> Repo.delete()
   end
 
   @doc """